DEVPROD-5331 Reorganize /hooks folder
