--- apps/spruce/src/hooks/types.ts
@@ -1,3 +1,0 @@
-export interface TabToIndexMap {
-  [key: string]: number;
-}
