+++ apps/spruce/src/hooks/useBreadcrumbRoot/useBreadcrumbRoot.test.tsx
@@ -3,9 +3,9 @@ import { MockedProvider } from "@apollo/client/testing";
 import { OtherUserQuery, OtherUserQueryVariables } from "gql/generated/types";
 import { getUserMock } from "gql/mocks/getUser";
 import { OTHER_USER } from "gql/queries";
-import { useBreadcrumbRoot } from "hooks";
 import { renderHook, waitFor } from "test_utils";
 import { ApolloMock } from "types/gql";
+import { useBreadcrumbRoot } from ".";
 
 const cache = new InMemoryCache({
   typePolicies: {
