+++ apps/spruce/src/hooks/useDisableSpawnExpirationCheckbox/useDisableSpawnExpirationCheckbox.test.tsx
@@ -9,7 +9,7 @@ import { getSpruceConfigMock } from "gql/mocks/getSpruceConfig";
 import { MY_VOLUMES, MY_HOSTS } from "gql/queries";
 import { renderHook } from "test_utils";
 import { ApolloMock } from "types/gql";
-import { useDisableSpawnExpirationCheckbox } from "..";
+import { useDisableSpawnExpirationCheckbox } from ".";
 
 // @ts-expect-error: FIXME. This comment was added by an automated script.
 const getProvider = (mocks) => {
