+++ apps/spruce/src/hooks/useFirstDistro/index.ts
