+++ apps/spruce/src/hooks/useNetworkStatus/index.ts
