+++ apps/spruce/src/hooks/useOnClickOutside/index.ts
