+++ apps/spruce/src/hooks/usePageVisibility/index.ts
